Move conversion to convert.py and rename
