+++ phylokit/__init__.py
@@ -2,14 +2,14 @@ from .balance import b1_index  # NOQA
 from .balance import b2_index  # NOQA
 from .balance import colless_index  # NOQA
 from .balance import sackin_index  # NOQA
-from .core import dataset_to_tskit  # NOQA
-from .core import tskit_to_dataset  # NOQA
+from .convert import from_tskit  # NOQA
+from .convert import to_tskit  # NOQA
 
 __all__ = [
     "sackin_index",
     "colless_index",
     "b1_index",
     "b2_index",
-    "tskit_to_dataset",
-    "dataset_to_tskit",
+    "from_tskit",
+    "to_tskit",
 ]
