+++ phylokit/core.py
@@ -2,8 +2,6 @@ import logging
 import os
 
 import numba
-import numpy as np
-import tskit
 import xarray
 
 logger = logging.getLogger(__name__)
@@ -43,7 +41,8 @@ def numba_njit(func, **kwargs):
         return func
 
 
-def create_tree_dataset(parent, time, left_child, right_sib, samples):
+# TODO add some defaults
+def create_tree_dataset(*, parent, time, left_child, right_sib, samples):
     data_vars = {
         "node_parent": ([DIM_NODE], parent),
         "node_time": ([DIM_NODE], time),
@@ -52,31 +51,3 @@ def create_tree_dataset(parent, time, left_child, right_sib, samples):
         "sample_node": ([DIM_SAMPLE], samples),
     }
     return xarray.Dataset(data_vars)
-
-
-def tskit_to_dataset(tree):
-    ts = tree.tree_sequence
-    # NOTE: we need to add an extra element here to keep the arrays in the nodes
-    # dimension the same length
-    time = np.append(ts.tables.nodes.time, np.inf)
-    return create_tree_dataset(
-        parent=tree.parent_array,
-        time=time,
-        left_child=tree.left_child_array,
-        right_sib=tree.right_sib_array,
-        samples=ts.samples(),
-    )
-
-
-def dataset_to_tskit(ds):
-    tables = tskit.TableCollection(1)
-    N = ds.sizes["nodes"] - 1
-    flags = np.zeros(N, dtype=np.uint32)
-    flags[ds.sample_node.to_numpy()] = tskit.NODE_IS_SAMPLE
-    tables.nodes.set_columns(flags=flags, time=ds.node_time[:-1])
-    # TODO do this more efficiently.
-    for u, parent in enumerate(ds.node_parent):
-        if parent != -1:
-            tables.edges.add_row(0, 1, int(parent), u)
-    tables.sort()
-    return tables.tree_sequence().first()
